Create mongodb_hostpath.yml
This .yml file defines a Deployment and a Service for a MongoDB database in a Kubernetes cluster. The Deployment, named mongodb, is deployed in the nodejs-unir namespace and uses the official MongoDB Docker image. It is configured to mount a persistent volume from the host (/data/mongodb) to the MongoDB data directory (/data/db) inside the container, ensuring that the database data is preserved even if the container is restarted. The file also defines a Service named mongo, which exposes port 27017 so that other Pods within the cluster can access the MongoDB database.