Create mongodb_pv_pvc.yaml
This .yml file defines a PersistentVolume (PV), a PersistentVolumeClaim (PVC), a Deployment, and a Service for a MongoDB database in a Kubernetes cluster.

The PersistentVolume, named mongodb-pv, is provisioned in the nodejs-unir namespace with a capacity of 5Gi, backed by an AWS Elastic Block Store (EBS) volume. The volume is identified by vol-0c85a1f0ba02eded7 and formatted with the ext4 file system.

The PersistentVolumeClaim, named mongodb-pvc, requests 5Gi of storage in the same namespace and matches the label type: amazonEBS to bind with the corresponding PersistentVolume (mongodb-pv).

The Deployment, named mongodb, deploys a MongoDB instance using the official MongoDB Docker image. The Deployment is configured with a Recreate strategy, meaning existing Pods are terminated before new ones are created during updates. The MongoDB container mounts the PersistentVolumeClaim (mongodb-pvc) at /data/db to ensure persistent storage for MongoDB data.

Finally, the Service, named mongo, exposes the MongoDB instance on port 27017 within the nodejs-unir namespace, making the database accessible to other Pods within the Kubernetes cluster.