Create nodejs.yaml
This .yml file defines a Deployment and a Service for a Node.js application in a Kubernetes cluster.

The Deployment, named nodejs, is deployed in the nodejs-unir namespace. It specifies two replicas, meaning there will be two instances of the Node.js application running concurrently. The Deployment uses the Docker image juanlubh/node-dockercompose:1.0.0. The application is exposed on port 3000 inside the container. The Deployment is configured with a label app: nodejs, which is used to match Pods managed by this Deployment.

The Service, named nodejs-lb, also in the nodejs-unir namespace, is of type LoadBalancer. This Service automatically creates an external IP address that distributes traffic across the two Node.js application instances. It listens on port 3000 and forwards incoming traffic to the same port on the selected Pods (app: nodejs). The LoadBalancer ensures that the application is accessible from outside the Kubernetes cluster.